use proper http response status codes, document possible response codes in todo.md, added decorator to validate incoming json (needs cleanup but works)

+++ app.py
@@ -1,8 +1,26 @@
+from functools import wraps
 import sqlite3
-from flask import Flask, render_template, g, url_for, json, Response, request
+import jsonschema
+from flask import Flask, render_template, g, url_for, json, Response, request, abort
 app = Flask(__name__)
 
 DATABASE = 'test.db'
+
+
+def json_validate(schema):
+	def decorator(f):
+		@wraps(f)
+		def decorated_function(*args, **kwargs):
+			print schema
+			print 'json_validate'
+			v = jsonschema.Draft3Validator(schema)
+			errors = [str(error) for error in v.iter_errors(request.json)]
+			if len(errors) > 0:
+				return JSONResponse(json.dumps({'errors':errors}), 400)
+			return f(*args, **kwargs)
+		return decorated_function
+	return decorator
+
 
 def query_db(query, *args):
 	r = g.db.execute(query, args)
@@ -28,6 +46,8 @@ def sitemap():
 	routes.sort(cmp=lambda a,b: cmp(a.split()[1], b.split()[1]))
 	return Response('sitemap:'+json.dumps(routes, indent=2), mimetype='text/plain')
 
+def JSONResponse(json, status=200):
+	return Response(json, status, mimetype='application/json')
 
 #
 # Companies API
@@ -38,34 +58,51 @@ def companies():
 	return json.dumps(companies)
 
 @app.route('/companies', methods=['POST'])
+@json_validate(
+	{
+		'type': 'object',
+		'additionalProperties': False,
+		'properties': {
+			'name': { 'type': 'string' }
+		}
+	}
+)
 def post_company():
 	g.db.execute('insert into companies (name) values (?)', (request.json['name'],))
 	company_id = g.db.execute('select last_insert_rowid()').fetchone()[0]
 	g.db.commit()
-	return json.dumps(company_id)
+	return JSONResponse(json.dumps(company_id), 201)
 
 @app.route('/companies/<company_id>')
 def company(company_id):
 	companies = query_db('select * from companies where id=?', company_id)
-	return json.dumps(companies[0])
+	if len(companies) == 0:
+		return abort(404)
+	return JSONResponse(json.dumps(companies[0]))
 
 @app.route('/companies/<company_id>', methods=['PUT'])
 def put_company(company_id):
+	exists = g.db.execute('select exists(select * from companies where id=?)', (company_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('update companies set name=? where id=?', (request.json['name'], company_id))
 	g.db.commit()
 	companies = query_db('select * from companies where id=?', company_id)
-	return json.dumps(companies[0])
+	return Response('', 204)
 
 @app.route('/companies/<company_id>', methods=['DELETE'])
 def delete_company(company_id):
+	exists = g.db.execute('select exists(select * from companies where id=?)', (company_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('delete from companies where id=?', (company_id,))
 	g.db.commit()
-	return 'success'
+	return Response('', 204)
 
 @app.route('/companies/<company_id>/projects')
 def company_projects(company_id):
 	projects = query_db('select * from projects where company_id=?', company_id)
-	return json.dumps(projects)
+	return JSONResponse(json.dumps(projects))
 
 
 #
@@ -74,37 +111,55 @@ def company_projects(company_id):
 @app.route('/projects')
 def projects():
 	projects = query_db('select * from projects')
-	return json.dumps(projects)
+	return JSONResponse(json.dumps(projects))
 
 @app.route('/projects', methods=['POST'])
+@json_validate(
+	{
+		'type': 'object',
+		'additionalProperties': False,
+		'properties': {
+			'name': { 'type': 'string' },
+			'company_id': { 'type': 'integer' }
+		}
+	}
+)
 def post_project():
 	g.db.execute('insert into projects (name, company_id) values (?, ?)', (request.json['name'],request.json['company_id']))
 	project_id = g.db.execute('select last_insert_rowid()').fetchone()[0]
 	g.db.commit()
-	return json.dumps(project_id)
+	return JSONResponse(json.dumps(project_id), 201)
 
 @app.route('/projects/<project_id>')
 def project(project_id):
 	projects = query_db('select * from projects where id = ?', project_id)
-	return json.dumps(projects[0])
+	if len(projects) == 0:
+		return abort(404)
+	return JSONResponse(json.dumps(projects[0]))
 
 @app.route('/projects/<project_id>', methods=['PUT'])
 def put_project(project_id):
+	exists = g.db.execute('select exists(select * from projects where id=?)', (project_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('update projects set name=? where id=?', (request.json['name'], project_id))
 	g.db.commit()
 	projects = query_db('select * from projects where id=?', project_id)
-	return json.dumps(projects[0])
+	return Response('', 204)
 
 @app.route('/projects/<project_id>', methods=['DELETE'])
 def delete_project(project_id):
+	exists = g.db.execute('select exists(select * from projects where id=?)', (project_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('delete from projects where id=?', (project_id,))
 	g.db.commit()
-	return 'success'
+	return Response('', 204)
 
 @app.route('/projects/<project_id>/tasks')
 def project_tasks(project_id):
 	tasks = query_db('select * from tasks where project_id = ?', project_id)
-	return json.dumps(tasks)
+	return JSONResponse(json.dumps(tasks))
 
 
 #
@@ -113,37 +168,55 @@ def project_tasks(project_id):
 @app.route('/tasks')
 def tasks():
 	tasks = query_db('select * from tasks')
-	return json.dumps(tasks)
+	return JSONResponse(json.dumps(tasks))
 
 @app.route('/tasks', methods=['POST'])
+@json_validate(
+	{
+		'type': 'object',
+		'additionalProperties': False,
+		'properties': {
+			'title': { 'type': 'string' },
+			'project_id': { 'type': 'integer' }
+		}
+	}
+)
 def post_task():
 	g.db.execute('insert into tasks (title, project_id, date_created) values (?, ?, current_timestamp)', (request.json['title'],request.json['project_id']))
 	task_id = g.db.execute('select last_insert_rowid()').fetchone()[0]
 	g.db.commit()
-	return json.dumps(task_id)
+	return JSONResponse(json.dumps(task_id), 201)
 
 @app.route('/tasks/<task_id>')
 def task(task_id):
 	tasks = query_db('select * from tasks where id=?', task_id)
-	return json.dumps(tasks[0])
+	if len(tasks) == 0:
+		return abort(404)
+	return JSONResponse(json.dumps(tasks[0]))
 
 @app.route('/tasks/<task_id>', methods=['PUT'])
 def put_task(task_id):
+	exists = g.db.execute('select exists(select * from tasks where id=?)', (task_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('update tasks set title=?, project_id=? where id=?', (request.json['title'], request.json['project_id'], task_id))
 	g.db.commit()
 	tasks = query_db('select * from tasks where id=?', task_id)
-	return json.dumps(tasks[0])
+	return Response('', 204)
 
 @app.route('/tasks/<task_id>', methods=['DELETE'])
 def delete_task(task_id):
+	exists = g.db.execute('select exists(select * from tasks where id=?)', (task_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('delete from tasks where id=?', (task_id,))
 	g.db.commit()
-	return 'success'
+	return Response('', 204)
 
 @app.route('/tasks/<task_id>/notes')
 def task_notes(task_id):
 	notes = query_db('select * from notes where task_id=?', task_id)
-	return json.dumps(notes)
+	return JSONResponse(json.dumps(notes))
 
 
 #
@@ -152,32 +225,50 @@ def task_notes(task_id):
 @app.route('/notes')
 def notes():
 	notes = query_db('select * from notes')
-	return json.dumps(notes)
+	return JSONResponse(json.dumps(notes))
 
 @app.route('/notes', methods=['POST'])
+@json_validate(
+	{
+		'type': 'object',
+		'additionalProperties': False,
+		'properties': {
+			'text': { 'type': 'string' },
+			'task_id': { 'type': 'task_id' }
+		}
+	}
+)
 def post_note():
 	g.db.execute('insert into notes (text, task_id) values (?, ?)', (request.json['text'],request.json['task_id']))
 	task_id = g.db.execute('select last_insert_rowid()').fetchone()[0]
 	g.db.commit()
-	return json.dumps(task_id)
+	return JSONResponse(json.dumps(task_id), 201)
 
 @app.route('/notes/<note_id>')
 def note(note_id):
 	notes = query_db('select * from notes where id=?', note_id)
-	return json.dumps(notes[0])
+	if len(notes) == 0:
+		return abort(404)
+	return JSONResponse(json.dumps(notes[0]))
 
 @app.route('/notes/<note_id>', methods=['PUT'])
 def put_note(note_id):
+	exists = g.db.execute('select exists(select * from notes where id=?)', (note_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('update notes set text=? where id=?', (request.json['text'], note_id))
 	g.db.commit()
 	notes = query_db('select * from notes where id=?', note_id)
-	return json.dumps(notes[0])
+	return Response('', 204)
 
 @app.route('/notes/<note_id>', methods=['DELETE'])
 def delete_note(note_id):
+	exists = g.db.execute('select exists(select * from notes where id=?)', (note_id,)).fetchone()[0]
+	if not exists:
+		return abort(404)
 	g.db.execute('delete from notes where id=?', (note_id,))
 	g.db.commit()
-	return 'success'
+	return Response('', 204)
 
 
 #
@@ -191,12 +282,12 @@ def users():
 @app.route('/users/<user_id>')
 def user(user_id):
 	users = query_db('select * from users where id=?', user_id)
-	return json.dumps(users[0])
+	return JSONResponse(json.dumps(users[0]))
 
 @app.route('/users/<user_id>/tasks')
 def user_tasks(user_id):
 	tasks = query_db('select * from tasks where owner_id=?', user_id)
-	return json.dumps(tasks)
+	return JSONResponse(json.dumps(tasks))
